+++ map-structure/sensors/src/pointcloud-map-sensor.cc
@@ -54,11 +54,11 @@ void PointCloudMapSensor::saveToYamlNodeImpl(YAML::Node* sensor_node) const {
   }
   if (fov_upper_angle_deg_) {
     (*sensor_node)[static_cast<std::string>(kYamlFieldNameFoVUpperAngle)] =
-        YAML::convert<uint16_t>::encode(fov_upper_angle_deg_);
+        YAML::convert<float>::encode(fov_upper_angle_deg_);
   }
   if (fov_lower_angle_deg_) {
     (*sensor_node)[static_cast<std::string>(kYamlFieldNameFoVLowerAngle)] =
-        YAML::convert<uint16_t>::encode(fov_lower_angle_deg_);
+        YAML::convert<float>::encode(fov_lower_angle_deg_);
   }
 }
